extract evaporate status into a separate file
