+++ codemod/output/EvaporateStatusEnum.ts
@@ -1,0 +1,21 @@
+export enum EVAPORATE_STATUS {
+  PENDING = 0,
+  EVAPORATING = 2,
+  COMPLETE = 3,
+  PAUSED = 4,
+  CANCELED = 5,
+  ERROR = 10,
+  ABORTED = 20,
+  PAUSING = 30
+}
+
+export const PAUSED_STATUSES = [
+  EVAPORATE_STATUS.PAUSED,
+  EVAPORATE_STATUS.PAUSING
+]
+
+export const ACTIVE_STATUSES = [
+  EVAPORATE_STATUS.PENDING,
+  EVAPORATE_STATUS.EVAPORATING,
+  EVAPORATE_STATUS.ERROR
+]
